Unit test for bug QFJ-188
The test i ended up writing is a bit of an anti-test. There's no easy to get into the data structures of the client  initiator code to verify that the right SSL context is created.

so instead, i supply a bogus password that causes an error during the client initiator start since the SSL context can't be initialized. That verifies that the code goes through the same codepath of using/initializing the keystoreManagers when creating the SSL context (which was the change) so i think that may be enough.